% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alexa.R
\name{alexa}
\alias{alexa}
\title{vcs UI}
\usage{
alexa(rootpath = getwd())
}
\arguments{
\item{rootpath}{root path to look for version control directories, Default: getwd()}
}
\value{
nothing
}
\description{
Shiny gadget for controlling, navigating and maintaining version control directories on local and remote repositories
}
\details{
Version control systems that can be accessed are git, github, bitbucket and svn. Throught this gadget you can do full and sparse checkout and update current vcs directory.
}
\examples{
\dontrun{
if(interactive()){
 alexa(getwd())
 }
}
}
\seealso{
\code{\link[jsonlite]{fromJSON}}
 \code{\link[jsTree]{jsTreeOutput}},\code{\link[jsTree]{renderJsTree}}
 \code{\link[miniUI]{miniPage}},\code{\link[miniUI]{gadgetTitleBar}},\code{\link[miniUI]{miniTitleBarButton}},\code{\link[miniUI]{miniContentPanel}}
 \code{\link[shiny]{sidebarLayout}},\code{\link[shiny]{sidebarPanel}},\code{\link[shiny]{radioButtons}},\code{\link[shiny]{conditionalPanel}},\code{\link[shiny]{wellPanel}},\code{\link[shiny]{textInput}},\code{\link[shiny]{uiOutput}},\code{\link[shiny]{actionButton}},\code{\link[shiny]{reactiveValues}},\code{\link[shiny]{observeEvent}},\code{\link[shiny]{renderPrint}},\code{\link[shiny]{eventReactive}},\code{\link[shiny]{isolate}},\code{\link[shiny]{renderUI}},\code{\link[shiny]{verbatimTextOutput}},\code{\link[shiny]{selectInput}},\code{\link[shiny]{stopApp}},\code{\link[shiny]{runGadget}},\code{\link[shiny]{paneViewer}}
 \code{\link[vcs]{find.remote}},\code{\link[vcs]{diff_head}},\code{\link[vcs]{navigate_remote}},\code{\link[vcs]{sparse_checkout}}
}
