% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_pipe.R
\name{collapse_pipe}
\alias{collapse_pipe}
\title{Collapse lines of piped calls to single lines}
\usage{
collapse_pipe(pattern = "\%[[:punct:]]\%|[+]", x,
  comment_tokens = NULL)
}
\arguments{
\item{pattern}{character, symbols to collapse on Default: '\%[[:punct:]]\%|[+]'}

\item{x}{character, lines of script to run on}

\item{comment_tokens}{character, tokens to comment out collapsed lines e.g. 'plot', 'print' Default: NULL}
}
\value{
character
}
\description{
When parsing lines of code this will collapse piped lines into a single one
}
\examples{
\dontrun{
 str1<-'iris\%>\%
 dplyr::mutate(Sepal.Length*2)\%>\%
 plot()'
 x<-strsplit(str1,'\\\\n')[[1]]
 collapse_pipe(x=x,comment_tokens = 'plot')
}
}
