% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_repos.R
\name{created.repos}
\alias{created.repos}
\title{Find Created Repositories of a Github User}
\usage{
created.repos(user)
}
\arguments{
\item{user}{character, Github user name}
}
\value{
data.frame
}
\description{
Find (or retrieve) details of all repositories created by a specified Github user
}
\examples{
\dontrun{
if(interactive()){
 created.repos('tidyverse')
 }
}
}
