% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_remote.R
\name{find.remote}
\alias{find.remote}
\title{list.dir for git and svn repositories}
\usage{
find.remote(path, vcs = c("git", "svn"))
}
\arguments{
\item{path}{character, path to seach for svn repositories}

\item{vcs}{character, vector of what vcs systems to look for, Default: 'git'}
}
\value{
character
}
\description{
list parent directories of svn repositores within a path
}
\examples{
\dontrun{
if(interactive()){
 find.remote('/data')
 }
}
}
