% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_remote.R
\name{grepr}
\alias{grepr}
\title{Recursive grep}
\usage{
grepr(pattern, path, recursive = FALSE, padding = 0, ...)
}
\arguments{
\item{pattern}{character, string containing a regular expression}

\item{path}{character, path to search, see details}

\item{recursive}{logical, Should the listing recurse into directories? passed to list.files, Default: FALSE}

\item{padding}{integer, number of rows to return in addition to the query line, Default: 5}

\item{...}{arguments passed to grep}
}
\value{
grepr(value = FALSE) returns a vector of the indices of the elements of x 
that yielded a match (or not, for invert = TRUE. 
This will be an integer vector unless the input is a long vector, when it will be a double vector.
grepr(value = TRUE) returns a character vector containing the selected elements of x 
(after coercion, preserving names but no other attributes).
}
\description{
Recursive call to grep in R
}
\details{
if path is a character then the pattern is checked against a location on the local machine, 
if path is a list with arguments containing params for ls_remote
then the pattern is checked on the specified remote repository (repository must be public).
In this case of the recursive parameter takes over-rides over the list parameter.
}
\examples{
grepr(pattern = 'gsub',path = '.',value=TRUE,recursive = T)
grepr(pattern = 'importFrom',path = list(path='yonicd/vcs',subdir='R'))
grepr(pattern = 'importFrom',path = list(path='yonicd/vcs',subdir='R'),value=TRUE)
grepr(pattern = 'importFrom',path = list(path='yonicd/vcs',subdir='R'),value=TRUE,padding=3)
grepr(pattern = 'tags$script',path = list(path = 'timelyportfolio/vueR',subdir='R|inst'),padding=3,value=TRUE,fixed=TRUE)
}
