% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_repos.R
\name{list_repos}
\alias{list_repos}
\title{Find Created Repositories of a Github User}
\usage{
list_repos(user, vcs = "github", fields = "full_name")
}
\arguments{
\item{user}{character, Github user name}

\item{vcs}{character, choose which version control system to search (github, bitbucket), Default: 'github'}

\item{fields}{character, fields of repository details, Default='full_name'}
}
\value{
list
}
\description{
Find (or retrieve) details of all repositories created by a specified Github user
}
\details{
The fields to choose from are detailed in the following link for github
 \url{https://developer.github.com/v3/search/#search-repositories}, for bitbucket 
 \url{https://developer.atlassian.com/bitbucket/api/2/reference/resource/repositories}
}
\examples{
\dontrun{
if(interactive()){
 list_repos('tidyverse',vcs='github',fields='full_name')
 list_repos('metrumrg',vcs='bitbucket',fields='full_name')
 }
}
}
