% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_github.R
\name{ls_github}
\alias{ls_github}
\title{List files of a github repository subdirectory}
\usage{
ls_github(repo, branch = "master", subdir = NULL, full.names = FALSE,
  recursive = FALSE)
}
\arguments{
\item{repo}{character, Repository address in the format username/repo}

\item{branch}{character, Branch of repo, Default: 'master'}

\item{subdir}{character, subdirectory within repo that contains the files to list}

\item{full.names}{boolean, If TRUE, the directory path is prepended to the file names to give a relative file path. 
If FALSE, the file names (rather than paths) are returned, Default: TRUE}

\item{recursive}{boolean, Should the listing recurse into directories?, Default: FALSE}
}
\description{
List full raw paths to files in a github repository subdirectory
}
\examples{
l=ls_github('tidyverse/ggplot2')
l
#geom-boxplot.r
readLines(l[42])
}
