% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_head.R
\name{ls_head}
\alias{ls_head}
\title{Inspect tree structure of repository head}
\usage{
ls_head(path = getwd(), branch = "master")
}
\arguments{
\item{path}{character, Path to root directory of git repository or a name of a github/bitbucket repository, Default: setwd()}

\item{branch}{character, alias of the github/bitbucket branch to retrieve directory structure, Default: 'master'}
}
\value{
data.frame
}
\description{
Inspect tree structure of git/bitbuck/svn repository head
}
\details{
By default path assumes a local address, if path is a valid repository name eg 'tidyverse/glue' 
then the pattern is checked on the specified pulic github repository
}
\examples{
ls_head('tidyverse/glue') #github ('master' branch)
ls_head('tidyverse/glue',branch='named_args') #github ('named_args' branch)
ls_head('metrumrg/qapply') #bitbucket ('master' branch)

}
