% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_remote.R
\name{ls_remote}
\alias{ls_remote}
\title{Inspect directory structure of version control system repository}
\usage{
ls_remote(
  path = getwd(),
  branch = "master",
  subdir = NULL,
  vcs = "github",
  full.names = FALSE,
  PAT = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{path}{character, Path to root directory of git/svn repository or a name of a github/bitbucket repository, Default: setwd()}

\item{branch}{character, alias of the branch in the repository, Default: 'master'}

\item{subdir}{chracter, subdirectory of repository/branch to query, if '.' then only the root directory will be queried, Default: NULL}

\item{vcs}{character, choose which version control system to search (github, bitbucket, git (head of checkout repo), svn), Default: 'github'}

\item{full.names}{boolean, If TRUE, the directory path is prepended to the file names to give a relative file path. 
If FALSE, the file names (rather than paths) are returned, Default: TRUE}
}
\value{
character
}
\description{
Inspect directory structure of git/bitbuck/svn repository head
}
\examples{
#github
  ls_remote('tidyverse/glue')
  #'named_args' branch
    ls_remote('tidyverse/glue',branch='named_args')
  #'named_args' branch only 'src' subdirectory
    ls_remote('tidyverse/glue',branch='named_args',subdir='src')
 #bitbucket
   #'master' branch
    ls_remote('metrumrg/qapply',vcs='bitbucket')

}
