% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_github.R
\name{setwd_github}
\alias{setwd_github}
\title{Set the working directory to a Github repository URL address}
\usage{
setwd_github(repo, subdir, r.file, add_keywords = "stan")
}
\arguments{
\item{path}{URL address}
}
\description{
Appends to read and source calls URL path.
}
