% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setwd_remote.R
\name{setwd_remote}
\alias{setwd_remote}
\title{Set the working directory to a Github repository URL address}
\usage{
setwd_remote(r.script.raw, repo, subdir, vcs = "github", tokens = NULL)
}
\arguments{
\item{r.script.raw}{character, script to append remote path.}

\item{tokens}{character, vectors of functions names that will be have a path appended to them.}

\item{path}{character, path to append}
}
\description{
Appends to read and source calls URL path.
}
