% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_github.R
\name{source_github}
\alias{source_github}
\title{FUNCTION_TITLE}
\usage{
source_github(repo, subdir, r.file, flag = TRUE)
}
\arguments{
\item{repo}{PARAM_DESCRIPTION}

\item{subdir}{PARAM_DESCRIPTION}

\item{r.file}{PARAM_DESCRIPTION}

\item{flag}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
repo='stan-dev/example-models'
subdir='ARM/Ch.10'
r.file='10.6_IVinaRegressionFramework.R'
source_github(repo,subdir,r.file)
}
