% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_github.R
\name{sparse_github}
\alias{sparse_github}
\title{Create a sparse git repository}
\usage{
sparse_github(repo_url, repo, dirs, create = TRUE, append = TRUE,
  remote = "origin", branch = "master")
}
\arguments{
\item{repo_url}{character, repository url path}

\item{repo}{character, path on local disk that repository is cloned to}

\item{dirs}{character, vector of repository subdirectories to fetch}

\item{create}{boolean, create a new git clone?, Default: TRUE}

\item{append}{boolean, append new lines to sparse-checkout file, Default: TRUE}

\item{remote}{character, alias of the remote Default: 'origin'}

\item{branch}{character, alias of the branch, Default: 'master'}
}
\value{
nothing
}
\description{
Create a git repository that fetches only specified subdirectories and file types
}
\examples{
\donttest{
repo_url='https://github.com/tidyverse/ggplot2.git'
repo='ggplot2-sparse'
dirs=c('data-raw/*.csv','man/*.Rd')
#create new sparse clone
sparse_github(repo_url,repo,dirs)

#update sparse-checkout definitions (appends to current list)
sparse_github(repo_url,repo,dirs=c('man/macros/*.Rd'),create=FALSE,append=TRUE)
}
}
