% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_depth.R
\name{update_depth}
\alias{update_depth}
\title{Update depth definitions of checked out repository}
\usage{
update_depth(path, action, action_input)
}
\arguments{
\item{path}{character, path on local disk that repository is cloned to}

\item{action}{character, apply one of 'deepen','shallow-since','shallow-exclude','unshallow'}

\item{action_input}{parameter needed for the action (see details)}
}
\value{
nothing
}
\description{
Update the settings of a repository that was checked out with a depth parameter
}
\details{
action_input for each action: [class of input]:
'deepen': integer
specifies the number of commits from the current shallow boundary instead of from the tip of each remote branch history.

'shallow-since': date (%Y-%m-%d)
 Deepen or shorten the history of a shallow repository to include all reachable commits after a given date.

'shallow-exclude': integer
 Deepen or shorten the history of a shallow repository to exclude commits reachable from a specified remote branch or tag.

'unshallow': boolean
 If the source repository is complete, convert a shallow repository to a complete one, removing all the limitations imposed by shallow repositories.
}
