% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigate_remote.R
\name{navigate_remote}
\alias{navigate_remote}
\title{Visually inspect structure of git repository before initial fetch}
\usage{
navigate_remote(path = getwd(), branch = "master", vcs = "github",
  show_tree = TRUE)
}
\arguments{
\item{path}{character, Path to root directory of git repository or a name of a github/bitbucket repository, Default: setwd()}

\item{branch}{character, alias of the github/bitbucket branch to retrieve directory structure, Default: 'master'}

\item{vcs}{character, choose which version control system to search (github, bitbucket, git (head of checkout repo), svn), Default: 'github'}

\item{show_tree}{boolean, show jsTree widget output in viewer, Default: TRUE}
}
\value{
widget
}
\description{
Visually inpsect the directory structure of it before cloning/fetching/pulling
}
\details{
By default path assumes a local address, if path is a valid repository name eg 'tidyverse/glue' 
then the pattern is checked on the specified pulic github repository
}
\examples{
navigate_remote('tidyverse/glue')
navigate_remote('tidyverse/glue',branch='named_args')
}
\seealso{
\code{\link[jsTree]{jsTree}}
}
