% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_remote.R
\name{source_remote}
\alias{source_remote}
\title{Source File on a Remote Repository}
\usage{
source_remote(repo, subdir, r.file, tokens = NULL, branch = "master",
  vcs = "github", flag = TRUE)
}
\arguments{
\item{repo}{character, username/repo_name}

\item{subdir}{character, subdirectory of repository}

\item{r.file}{character, file name to source}

\item{tokens}{character, vector of functions names that will be have a path appended to them}

\item{branch}{character, alias of the branch in the repository, Default: 'master'}

\item{vcs}{character, choose which version control system to search (github, bitbucket), Default: 'github'}

\item{flag}{boolean, checks to see if the file is a nested file Default: TRUE}
}
\description{
Evaluate script directly from remote repository 
including reading data and nested sources.
}
\examples{
\donttest{
repo='stan-dev/example-models'
subdir='ARM/Ch.10'
r.file='10.4_LackOfOverlapWhenTreat.AssignmentIsUnknown.R'
source_remote(repo,subdir,r.file,tokens='stan')
}
}
